update of error validation configuration, organization and commissioning
