--- src/main/java/com/sami/sami_app/api/controllers/BadRequestController.java
@@ -1,43 +1,0 @@
-package com.sami.sami_app.api.controllers;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import com.sami.sami_app.api.error_handler.BaseErrorResponse;
-import com.sami.sami_app.api.error_handler.ErrorResponse;
-import com.sami.sami_app.api.error_handler.ListErrorsResponse;
-import com.sami.sami_app.util.exceptions.IdNotFoundException;
-
-@RestControllerAdvice
-@ResponseStatus(code = HttpStatus.BAD_REQUEST)
-public class BadRequestController {
-
-    @ExceptionHandler(IdNotFoundException.class)
-    public BaseErrorResponse handleIdNotFuound(IdNotFoundException exception) {
-
-        return ErrorResponse.builder()
-                .code(HttpStatus.BAD_REQUEST.value())
-                .status(HttpStatus.BAD_REQUEST.name())
-                .message(exception.getMessage()).build();
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public BaseErrorResponse handleErrors(MethodArgumentNotValidException exception) {
-        List<String> errors = new ArrayList<>();
-
-        exception.getAllErrors().forEach(error -> errors.add(error.getDefaultMessage()));
-
-        return ListErrorsResponse.builder()
-                .code(HttpStatus.BAD_REQUEST.value())
-                .status(HttpStatus.BAD_REQUEST.name())
-                .errors(errors)
-                .build();
-    }
-
-}
