+++ src/main/java/com/sami/sami_app/api/error_handler/BadRequestController.java
@@ -1,0 +1,58 @@
+package com.sami.sami_app.api.error_handler;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import com.sami.sami_app.api.dto.errors.BaseErrorResponse;
+import com.sami.sami_app.api.dto.errors.ListErrorsResponse;
+import com.sami.sami_app.util.exceptions.BadRequestException;
+
+@RestControllerAdvice
+@ResponseStatus(code = HttpStatus.BAD_REQUEST)
+public class BadRequestController {
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public BaseErrorResponse handleErrors(MethodArgumentNotValidException exception) {
+
+        List<Map<String, String>> errors = new ArrayList<>();
+
+        exception.getBindingResult().getFieldErrors().forEach(e -> {
+            Map<String, String> error = new HashMap<>();
+            error.put("error", e.getDefaultMessage());
+            error.put("field", e.getField());
+            errors.add(error);
+        });
+
+        return ListErrorsResponse.builder()
+                .code(HttpStatus.BAD_REQUEST.value())
+                .status(HttpStatus.BAD_REQUEST.name())
+                .errors(errors)
+                .build();
+    }
+
+    @ExceptionHandler(BadRequestException.class)
+    public BaseErrorResponse handleError(BadRequestException exception) {
+        List<Map<String, String>> errors = new ArrayList<>();
+
+        Map<String, String> error = new HashMap<>();
+
+        error.put("ID", exception.getMessage());
+
+        errors.add(error);
+
+        return ListErrorsResponse.builder()
+                .code(HttpStatus.BAD_REQUEST.value())
+                .status(HttpStatus.BAD_REQUEST.name())
+                .errors(errors)
+                .build();
+
+    }
+}
