--- src/main/java/com/sami/sami_app/api/error_handler/ErrorResponse.java
@@ -1,17 +1,0 @@
-package com.sami.sami_app.api.error_handler;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-@SuperBuilder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ErrorResponse extends BaseErrorResponse {
-    private String message;
-
-}
