+++ src/main/java/com/sami/sami_app/util/exceptions/BadRequestException.java
@@ -1,0 +1,8 @@
+package com.sami.sami_app.util.exceptions;
+
+public class BadRequestException extends RuntimeException {
+    
+    public BadRequestException(String message){
+        super(message);
+    }
+}
